+++ tests/test_schema.py
@@ -1789,7 +1789,7 @@ class MySchema(Schema):
     email = fields.Email()
     age = fields.Integer()
 
-    def handle_error(self, errors, obj):
+    def handle_error(self, errors, obj, **kwargs):
         raise CustomError("Something bad happened")
 
 
@@ -1805,7 +1805,7 @@ class TestHandleError:
         class MySchema3(Schema):
             email = fields.Email()
 
-            def handle_error(self, error, data):
+            def handle_error(self, error, data, **kwargs):
                 assert type(error) is ValidationError
                 assert "email" in error.messages
                 assert list(error.messages.keys()) == ["email"]
@@ -1822,7 +1822,7 @@ class TestHandleError:
             email = fields.Email()
             url = fields.URL()
 
-            def handle_error(self, error, data):
+            def handle_error(self, error, data, **kwargs):
                 assert type(error) is ValidationError
                 assert "email" in error.messages
                 assert list(error.messages.keys()) == ["email"]
@@ -1843,7 +1843,7 @@ class TestHandleError:
                 if value < 0:
                     raise ValidationError("Must be greater than 0.")
 
-            def handle_error(self, error, data):
+            def handle_error(self, error, data, **kwargs):
                 assert type(error) is ValidationError
                 assert "num" in error.messages
                 assert list(error.messages.keys()) == ["num"]
@@ -1863,7 +1863,7 @@ class TestHandleError:
             def validates_schema(self, data, **kwargs):
                 raise ValidationError("Invalid schema!")
 
-            def handle_error(self, error, data):
+            def handle_error(self, error, data, **kwargs):
                 assert type(error) is ValidationError
                 assert list(error.messages.keys()) == ["_schema"]
                 assert data == in_data